+++ dot15d4/src/frame/ie/mod.rs
@@ -62,11 +62,3 @@ impl<T: AsRef<[u8]>> InformationElements<T> {
     }
 }
 
-/// A high-level representation of Information Elements.
-#[derive(Debug)]
-pub struct InformationElementsRepr {
-    /// The header information elements.
-    pub header_information_elements: Vec<HeaderInformationElementRepr, 16>,
-    /// The payload information elements.
-    pub payload_information_elements: Vec<PayloadInformationElementRepr, 16>,
-}
