+++ dot15d4/src/frame/ie/payloads.rs
@@ -1,4 +1,4 @@
-use super::{NestedInformationElementRepr, NestedInformationElementsIterator};
+use super::NestedInformationElementsIterator;
 use heapless::Vec;
 
 /// A reader/writer for the IEEE 802.15.4 Payload Information Elements.
@@ -118,30 +118,6 @@ impl<'f> Iterator for PayloadInformationElementsIterator<'f> {
             }
 
             Some(ie)
-        }
-    }
-}
-
-/// A high-level representation of a Payload Information Element.
-#[derive(Debug)]
-pub enum PayloadInformationElementRepr {
-    Mlme(Vec<NestedInformationElementRepr, 16>),
-}
-
-impl PayloadInformationElementRepr {
-    pub fn parse(ie: PayloadInformationElement<&[u8]>) -> Self {
-        match ie.group_id() {
-            PayloadGroupId::Mlme => {
-                let mut nested_information_elements = Vec::new();
-
-                for nested_ie in ie.nested_information_elements() {
-                    nested_information_elements
-                        .push(NestedInformationElementRepr::parse(nested_ie));
-                }
-
-                Self::Mlme(nested_information_elements)
-            }
-            _ => todo!(),
         }
     }
 }
